services/admin-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
uuid6>=2024.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis==5.0.4
alembic==1.13.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.22
structlog>=24.1.0
prometheus-client>=0.20.0
